refactor: GameController 안에서 경주하고 결과 보여줌
GameControll 클래스 안에 repeactRace 함수 생성
showRaceResult 함수 View클래스 내부로 이동
