+++ src/game/GameController.js
@@ -1,4 +1,5 @@
 import Car from './Car.js';
+import View from '../view/View.js';
 
 function getCars(userInput) {
   const cars = [];
@@ -9,8 +10,21 @@ function getCars(userInput) {
   return cars;
 }
 
+function goOnce(cars) {
+  cars.forEach((car) => car.go());
+}
+
 export default class GameController {
   constructor(userInput) {
-    this.cars = getCars(userInput);
+    this.userInput = userInput;
+    this.view = new View();
+    this.cars = getCars(this.userInput);
+  }
+
+  repeatRace() {
+    for (let index = 0; index < this.userInput.getRepeatCount(); index += 1) {
+      goOnce(this.cars);
+      this.view.showRaceResult(this.cars);
+    }
   }
 }
